requirements.txt
pandas
seaborn
streamlit
watchdog
streamlit-drawable-canvas
azure-storage-blob